api/requirements.txt
fastapi>=0.109.2
uvicorn[standard]>=0.27.1
pydantic>=2.0.0
pydantic-settings>=2.1.0
httpx>=0.26.0
redis>=5.0.1
python-dotenv>=1.0.1
google-genai>=0.3.0
tenacity>=8.2.3
orjson>=3.9.13
structlog>=24.1.0
groq>=0.4.2
markdown>=3.5.2
pytest>=8.2.0
pytest-asyncio>=0.24.0
pytest-mock>=3.12.0

